♻️ Make astro header component
Changes:
1. Change header component to use Astro instead of Vue.
2. Refactor existing components to make header easier to understand.

Test plan: Expect header bar functionality to be unchanged. Regression test:
1. Click Centrapay logo, redirected home.
2. Click Docs, redirected home.
3. Click search, can enter term and click link.
4. Click '/' to open command palette, can then press escape to close.
5. Mobile: Can open mobile menu and click link.
