+++ src/components/AppShell.astro
@@ -1,5 +1,5 @@
 ---
-import Header from './Header.vue';
+import Header from './Header.astro';
 import DesktopNavigation from './DesktopNavigation.astro';
 const { path, navigation } = Astro.props;
 
@@ -7,7 +7,7 @@ const navPath = path.endsWith('/') ? path.slice(0, -1) : path;
 ---
 
 <div class="flex h-screen w-screen">
-  <Header client:load path={navPath} navigation={navigation} />
+  <Header path={navPath} navigation={navigation} />
   <DesktopNavigation path={navPath} navigation={navigation} />
   <main class="flex flex-col min-h-full w-full pt-16 md:pl-64 xl:pl-80">
     <slot />
