+++ src/components/icons/CentrapayLogo.astro
@@ -1,0 +1,17 @@
+---
+const { class: className } = Astro.props;
+---
+<svg
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  xmlns="http://www.w3.org/2000/svg"
+  class:list={className}
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    fill="currentColor"
+    d="M12 3H14.25C17.9559 3 20.9689 6.0625 21 9.75C21 13.5 17.987 16.5 14.25 16.5H12V12.0026H9.75C8.50433 12.0026 7.46886 13.0058 7.5 14.25C7.5 15.4942 8.50433 16.5 9.75 16.5H12V20.9922L9.75 21C6.01298 21 3 17.9827 3 14.25C3 10.5173 6.01298 7.5 9.75 7.5H12V12H14.25C15.4957 12 16.4998 11 16.4998 9.75C16.4998 8.5 15.4957 7.5 14.25 7.5H12V3Z"
+  />
+</svg>
