+++ src/components/icons/CloseOutline.astro
@@ -1,0 +1,26 @@
+---
+const { class: className } = Astro.props;
+---
+<svg
+  width="24"
+  height="24"
+  viewBox="0 0 24 24"
+  fill="none"
+  aria-hidden="true"
+  xmlns:xlink="http://www.w3.org/1999/xlink"
+  xmlns="http://www.w3.org/2000/svg"
+  class:list={className}
+>
+  <path
+    d="M17.25 17.25L6.75 6.75"
+    stroke="currentColor"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+  />
+  <path
+    d="M17.25 6.75L6.75 17.25"
+    stroke="currentColor"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+  />
+</svg>
